Rename `sysroot` to `toolchain`
